- modified the table comparison function in totem.assert to compare tables recursively (tensors and storages are then compared for the max difference of their elements within tolerance), numbers are compared for equality within tolerance, other lua types are compared for equality, other userdata types raise an error
- modified the eq function in totem.Tester to call the corresponding
functions of assert, comparing storages, tensors or tables
respectively. Consequently, removed the functions _eqTable,
_eqStorage, _eqSize from totem.tester.

- the function eq should be the main entry point for testing as it can
compare any data types. The function assertTableEq now raises an
error if it is not given tables as arguments. The symmetry between
Tables, Tensors and Storages is increased (with the exception that
assertTableEqNew has to be used instead of assertTableEq, as the
later was already defined with different parameters).