Generic comparison is done by ensuring that the types of the two objects match and then applying the following rules. Table comparison is done element-wise (keys and types of the associated values must match and are then compared recursively). Torch tensors and storages are compared element-wise. Other types are compared with the regular Lua == operator (Lua numbers can be compared with a certain tolerance). Type-specific access to each comparison are also provided.
Changes implied for the user: table comparison is more generic (it used
to check for strict identity between tables), generic comparison and
tensors comparison is more specific ;comparing tensors of non-matching
types used to lead to erratic behavior -- either program crash
(comparing float with doubles with the specific tensor comparison) or
silent casting (comparing doubles with bytes or using the generic
equality comparison) -- the behavior is now that any type mismatched
lead to the failure of the equality test. It is the user responsibility
to do any casting before comparing the objects.

Changes implied in the code: added storage comparison, re-factorized a
bit the tensor comparison, modified the table comparison, re-factorized
a bit the general equality comparison. Some code cleaning here and
there. Some added tests and examples. Removed some tests that ensured
that no error-message was returned on success